requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
pytesseract>=0.3.10
Pillow>=9.0.0
numpy>=1.22.0
PyYAML>=6.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0
pandas>=1.4.0
scikit-learn>=1.1.0
imutils>=0.5.4
